backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector
requests
python-dotenv
